Group meetings e2e tests by user-facing capability.
Replace CRUD and ticket-based names with Schedule, Edit, Participants,
Delete, Notifications, and Call, and keep all cases in one spec.
